refactor: extract CreateData and EditData
